fix(core): consolidate scanning
